fix(challenge-detail): data가 undefined일 때 처리 / API 코드 수정, 응답 데이터 타입 모듈화
