+++ src/pages/challenge-detail/index.tsx
@@ -1,42 +1,19 @@
-import { useState, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
 import { getChallengeDetail } from '../../apis/challenge-detail/challenge.detail.api';
 import Description from './description/';
 import Ranking from './ranking/';
 import {
-  TabsContainer,
-  TabPanelContainer,
   Image,
   ImageMask,
+  TabPanelContainer,
+  TabsContainer,
   Wrapper,
 } from './styles';
+import { type ChallengeDetailData } from '@/apis/challenge-detail/challenge.detail.response';
 import DefaultImage from '@/assets/Default-Image.svg';
-import { Tab, Tabs, TabPanel } from '@/components/common/tab';
+import { Tab, TabPanel, Tabs } from '@/components/common/tab';
 import { TopBar } from '@/components/common/top-bar';
-
-type Challenge = {
-  id: number;
-  participantCount: number;
-  difficulty: number;
-  onceExp: number;
-  successExp: number;
-  dayType: 'DAY' | 'WEEK' | 'MONTH' | 'YEAR';
-  dayCount: number;
-};
-
-export type ChallengeDetailData = {
-  id: number;
-  title: string;
-  content: string;
-  participantCount: number;
-  startDate: string;
-  endDate: string;
-  category: 'HEALTH' | 'ECHO' | 'SHARE' | 'VOLUNTEER' | 'ETC';
-  guide: string;
-  maxDifficulty: number;
-  imageUrls: string[];
-  challenges: Challenge[];
-};
 
 export const ChallengeDetailPage = () => {
   const tabsList = [
@@ -44,17 +21,24 @@ export const ChallengeDetailPage = () => {
     { label: '랭킹', value: 1 },
   ];
   const [activeTab, setActiveTab] = useState<number>(0);
-  const [data, setData] = useState<ChallengeDetailData | null>(null);
+  const [data, setData] = useState<ChallengeDetailData | undefined>(undefined);
 
   const handleSelectedTab = (value: number) => {
     setActiveTab(value);
   };
 
   useEffect(() => {
-    getChallengeDetail(1).then((res) => {
-      setData(res);
-      console.log(res);
-    });
+    const fetchChallengeDetail = async () => {
+      try {
+        const res = await getChallengeDetail(1);
+        setData(res);
+        console.log(res);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+
+    fetchChallengeDetail();
   }, []);
 
   return (
